+++ parity-match/src/test/java/org/jvirtanen/parity/match/MarketEvents.java
@@ -18,18 +18,18 @@ class MarketEvents implements MarketListener {
 
     @Override
     public void match(long restingOrderId, long incomingOrderId, long price,
-            int executedQuantity, int remainingQuantity) {
+            long executedQuantity, long remainingQuantity) {
         events.add(new Match(restingOrderId, incomingOrderId, price,
                     executedQuantity, remainingQuantity));
     }
 
     @Override
-    public void add(long orderId, Side side, long price, int size) {
+    public void add(long orderId, Side side, long price, long size) {
         events.add(new Add(orderId, side, price, size));
     }
 
     @Override
-    public void cancel(long orderId, int canceledQuantity, int remainingQuantity) {
+    public void cancel(long orderId, long canceledQuantity, long remainingQuantity) {
         events.add(new Cancel(orderId, canceledQuantity, remainingQuantity));
     }
 
@@ -40,11 +40,11 @@ class MarketEvents implements MarketListener {
         public final long restingOrderId;
         public final long incomingOrderId;
         public final long price;
-        public final int  executedQuantity;
-        public final int  remainingQuantity;
+        public final long executedQuantity;
+        public final long remainingQuantity;
 
         public Match(long restingOrderId, long incomingOrderId, long price,
-                int executedQuantity, int remainingQuantity) {
+                long executedQuantity, long remainingQuantity) {
             this.restingOrderId    = restingOrderId;
             this.incomingOrderId   = incomingOrderId;
             this.price             = price;
@@ -57,9 +57,9 @@ class MarketEvents implements MarketListener {
         public final long orderId;
         public final Side side;
         public final long price;
-        public final int  size;
+        public final long size;
 
-        public Add(long orderId, Side side, long price, int size) {
+        public Add(long orderId, Side side, long price, long size) {
             this.orderId = orderId;
             this.side    = side;
             this.price   = price;
@@ -69,10 +69,10 @@ class MarketEvents implements MarketListener {
 
     public static class Cancel extends Value implements Event {
         public final long orderId;
-        public final int  canceledQuantity;
-        public final int  remainingQuantity;
+        public final long canceledQuantity;
+        public final long remainingQuantity;
 
-        public Cancel(long orderId, int canceledQuantity, int remainingQuantity) {
+        public Cancel(long orderId, long canceledQuantity, long remainingQuantity) {
             this.orderId           = orderId;
             this.canceledQuantity  = canceledQuantity;
             this.remainingQuantity = remainingQuantity;
